requirements.txt
deepagents>=0.4.1
tavily-python>=0.5.0
streamlit>=1.40.0
python-dotenv>=1.0.0
langchain-openai>=0.3.0